+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Unreleased
 
 # 0.2.1
+- Added `bounds!` macro for `where` clauses on generic dimensions
+- Fixed documentation
 
 # 0.2.0
 - There are now `Tensor` instead of `Matrix`